updated logic and added test MOQ

+++ eBidder/Services/UserService.cs
@@ -29,14 +29,31 @@ namespace eBidder.Services
 
         public UserViewModel GetByUsername(string username)
         {
+            if (username == null)
+            {
+                throw new ArgumentNullException("Username must be provided");
+            }
+
             return _userRepository.GetByUsername(username)?.ToUserViewModel();
         }
 
         public UserViewModel CreateUser(string username, string password)
         {
-            if(_userRepository.GetByUsername(username) != null)
+            if (username == null)
             {
-                throw new ArgumentException($"User with given username: {username} already exists.");
+                throw new ArgumentNullException("Username must be provided");
+            }
+
+            if (password == null)
+            {
+                throw new ArgumentNullException("Password must be provided");
+            }
+
+            var existingUser = GetUser(username);
+
+            if (existingUser != null)
+            {
+                throw new InvalidOperationException($"User {username} already exists");
             }
 
             return _userRepository.CreateUser(username, password).ToUserViewModel();
@@ -44,12 +61,55 @@ namespace eBidder.Services
 
         public bool DeleteUser(string username)
         {
+            if (username == null)
+            {
+                throw new ArgumentNullException("Username must be provided");
+            }
+
+            var existingUser = GetUser(username);
+
+            if (existingUser == null)
+            {
+                return false;
+            }
+
             return _userRepository.DeleteUser(username);
         }
 
         public UserViewModel ChangePassword(string username, string oldPassword, string newPassword)
         {
+            if (username == null)
+            {
+                throw new ArgumentNullException("Username must be provided");
+            }
+
+            if (oldPassword == null)
+            {
+                throw new ArgumentNullException("OldPassword must be provided");
+            }
+
+            if (newPassword == null)
+            {
+                throw new ArgumentNullException("NewPassword must be provided");
+            }
+
+            var existingUser = GetUser(username);
+            if (existingUser == null)
+            {
+                throw new InvalidOperationException($"User {username} doesn't exist");
+            }
+
+            if (!existingUser.Password.Equals(oldPassword))
+            {
+                throw new ArgumentException("Wrong password for the user");
+            }
+
             return _userRepository.ChangePassword(username, oldPassword, newPassword).ToUserViewModel();
+        }
+
+        private UserViewModel GetUser(string username)
+        {
+            return _userRepository.GetByUsername(username)?.ToUserViewModel();
         }
     }
 }
